New browserify module
It has different API then older modules.

It exports a factory function that takes options and returns gulp task function.
Doesn't register a new task.

That way it is possible to use it in more then one task with different configuration.

In the future all modules will work this way.

Also added depreciation notice to coffeeify.